릴리즈 모드 오류 체크, 픽셀 충돌(벽) 충돌 테스트

+++ DX_HollowKnight/Contents/Particle.cpp
@@ -87,7 +87,7 @@ void AParticle::SpawnParticle()
 		Effect->SetZSort(EZOrder::HitParticleEffect);
 		Effect->SetScale(RandomScale);
 		Effect->GetRenderer()->SetAlpha(Alpha);
-		Effect->SetZSort(static_cast<int>(EZOrder::HitParticleEffect) -( i + 1 + ZOrderOffset));
+		Effect->SetZSort(static_cast<int>(EZOrder::HitParticleEffect) - ( i + 1 + static_cast<int>(ZOrderOffset)));
 
 		FParticleData NewParticle(Effect, Velocities[i]);
